Updates to Gradle to match previous Eclipse settings.
Note that once the transition to Gradle is complete, Eclipse project files should be removed from version control and generated by Gradle instead.

This commit downgrades to GDX 0.9.8 from the snapshot, and adds dependencies on the native jars. As GDX was not in the Sonatype repo for 0.9.8, these have been added to our own Artifactory repo.

Also added tasks runServer (in deco2800.arcade.server) and runArcade (in deco2800.arcade)