Add classes and handlers for secure handshake
Warning: the classes here have not been tested and some work is still
required to complete the handshake process.

Adds bunch of classes that help with the handshake between client and
server that establishes a shared secret key between the two parties.

Adds proxy listener that handles unsealing messages before the message
gets to other listeners.

Changes the send method of NetworkClient to seal messages with the
session key prior to transmission [more work to be done].

Now connection listeners actually go through the handshake process.