Working on achievement screen
I have created the screen, there is an achievement title at the top,
looping through all achievements and drawing them.

Achievement API is behind with this branch, so to fix that we have to
merge master again.

I placed Zoeys images into the serverResources directory, so that is
being saved on the DB :D YAY!