Added Deck, Player, User classes, Table Update
Deck and Player classes by Scott - Deck complete, Player still WIP
User class is a small container sort of class at the moment
Table class has most of the smaller functions complete, many of the main
game state functions need to be finished but contain working code.
This commit is mainly to show progress and will be working on this
further, hopefully to some level of near-completion, in the next few
days.