Store Overhaul and Relocation
Split the Store off from UI branch, and made interface.
Basically the Store is no longer a single page, but a fully functioning
interface, with methods, and planning and thought behind it.
Did huge UI overhaul for Store Main page, ready to be duplicated for all
4-5 store pages.

Looks like Java's finally realized that we don't have to be enemies. We
can co-operate, and get along, move forward for a mutual goal, through
understanding and experience.