Reformatted Mover.checkNoWallCollision to call Mover.nextTile and started implementing extra Dir variable.
The plan is to have two 'facing' variables; one is the actual drawn facing of pacman, the other is the direction
pacman is moving in. At the moment these are called 'drawFacing' and 'facing'. Perhaps it would be less confusing to later
refactor to more meaningfful names. The idea here is that 'drawFacing' only updates when
the tile in direction 'facing' is not a WallTile.

NOTE: NEED TO GET COLLISION BACK TO WORKING NORMALLY.

e.g. Left key is pressed along a horizontal corridor. 'drawFacing' and 'facing' are both LEFT.
Along this corridor, Up is pressed. Now 'drawFacing' stays LEFT and 'facing' changes to UP.
As soon as UP becomes a valid direction, 'drawFacing' updates to UP.