Store Faster, Fix Star Ratings, Store has Player
The store now takes a Player as an arguments, so it knows who is logged
in, and can perform actions with their info. This is yet to be put to
use, since the arcadeUI Player ONLY has an ID...

The star rating on the StoreGame page are fixed, and look wunnerful.

Made the entire Store really fast, by pausing each page instead of
re-creating them each time.